Assignment 2 Code
This is a basic pre-trained model used to classify a new image-dataset into two categories, namely "Interior" and "Dish" depending on whether the respective image displays a dish or is a photo of the restaurant's interior or something else.
The code does not yet save the model to disk and retrieves it upon training (something I'll work on next), meaning everytime you run this code, the model will be trained anew. Note that you of course need to download the training data and set up an appropriate directory/ folder-structure for all of this to work.